IXUS 980/ SD 990 beta disable fi2 for sx1, viewport_width for 990 merge
git-svn-id: https://tools.assembla.com/svn/chdk/trunk@740 6794e30b-3f2a-0410-a806-a2bbca1c07ff

+++ platform/sx1/sub/200h/lib.c
@@ -46,10 +46,13 @@ long vid_get_bitmap_screen_height()
 	return 240;
 }
 
-long vid_get_viewport_width()
+// default is same as bitmap
+/*
+int vid_get_viewport_width()
 {
 	return 480;
 }
+*/
 long vid_get_viewport_height()
 {
 	return 270;
